+++ CHANGELOG.md
@@ -2,7 +2,7 @@ CHANGELOG
 =========
 
 ## HEAD (Unreleased)
-_(none)_
+* Fix: transform "Example keyword" as `example`
 
 ---
 
@@ -13,12 +13,12 @@ _(none)_
 * Feature: Improved `scenarioOutline` syntax
 * Feature: Transform uses improved `scenarioOutline` syntax
 * Feature: Support `gherkin.skip` and `gherkin.only`
-* Feature: Even more improved Scenario Outline syntax
+* Feature: Even more improved "Scenario Outline" syntax
 
 ## 2.0.1 (2020-08-22)
 * Patch: Fix bin definition in package
 
 ## 2.0.0 (2020-08-22)
-* Breaking: Improved Scenario Outline syntax
+* Breaking: Improved "Scenario Outline" syntax
 * Feature: Add transform CLI